Move Report interface to 3_0 only
